refactor: extract DebugSessionMetadata.swift from DebugBundleTypes.swift
Splits DebugBundleTypes.swift into concern-focused files:
- DebugBundleTypes.swift keeps the snapshot value types
  (DebugInfoSnapshot data + DebugBundleSnapshot aggregate)
- DebugSessionMetadata.swift gets the large Codable session-metadata
  payload (~100 lines, includes recording/transcription/tracker
  configuration serialization)

Byte-identical move of lines 53-156. Both files import Foundation.

DebugBundleTypes.swift: 161 → 56 lines. Tests: 667.
